🩹 Fix find_*_of implementations
A bug was opened due to warnings and issues with the find_last_*
implementations. This has been fixed by refactoring all of them to use
a common "is_one_of" function for better readability, and unit tests
have been added to check for correctness in behavior.

Fixes bitwizeshift/string_view-standlone#2