✨ Add configuration header
This project aims to unify C++14, C++17, and C++20 types into C++11,
while still maintaining forward compatibility with the later versions
as well -- including with `constexpr` support.

The initial definition of C++11's constexpr is extremely constraining,
since it requires `constexpr` to only be on const-member functions,
and for each function to be exactly 1 line in length. This makes
defining conditional logic complicated without some way of referring
to the more "modern" C++ constexpr support.

To make this possible, a configuration header has been added that
checkes the current compilation model to determine how we are compiling
in order to see what support we currently have. This allows definitions
to be "constexpr" when the language supports it, but not constexpr if
using an older version like C++11.