✨ Add memory header
C++11 introduced the shared_ptr and unique_ptr types, along with the
convenience function make_shared -- however there was a notable
omission of a make_unique function. This omission was rectified in
C++14 by the introduction of the type.

This commit adds in the <memory> header which also introduces the same
make_unique function.

Additionally it also introduces the new C++20 make_unique_for_overwrite
function as well.