✨ Add <optional> header
C++17 introduced a new vocabulary type called 'optional', which
represents an optional value-type object. That is; an object which
may or may not contain a value, but does not exist through indirection
(e.g. like a pointer).

This type becomes extremely useful for conveying optional semantics on
APIs that may contain ownership; unlike the use of pointers, where
ownership may be ambiguous.

This commit introduces this header into Backport.