✨ Add to_address
C++17 introduced a new way to convert iterators and pointer-like types
to raw-pointers by using 'to_address'. This utility allows containers
to use generic logic with iterators so that contiguous iterators may
be converted to raw pointers in a coherent and consistent way.

This utility has been ported to C++11 in this library.