📇 Update feature_request github template
The feature_request github template was initially copied from an
example repo, and the contents of the file are not particularly
well-suited to this repository. This project does not actually
accept "feature requests" in the traditional sense, since the features
must be approved C++ standard library types.

As such, the template has been updated to be more specific so that
it explicitly outlines such cases.