🗑️ Remove non-standard extension from span
The span type was implemented with both 'iterator' and 'const_iterator'
member types; the latter of which is a non-standard extension to the
span interface. The C++ standard strangely only dictates 'iterator',
which may or may not be mutable based on whether T is mutable.

This design is... confusing, to say the least, which is how the
extensions came in in the first place -- since it was expected to
have feature parity with every other container wrapper out there.

The extensions have been removed.