✨ Add proxy_iterator type
The C++ standard refers to 'iterators' in a conceptual sense, never
referring specifically to an actual type -- but only in terms of
capability. In practice, this allows iterators to be just raw
pointers, since pointers may be iterated over. However, using
raw pointers as iterators in a container may be problematic, since
containers with similar underlying types may allow for their iterators
to be exchanged with one another -- something that is unsafe and
unwise to do.

To rectify this, a new type called 'proxy_iterator' has been added
as a detail type. This simply proxies the underlying type (e.g. a
pointer) but creates a new, unique type out of it.

This has been used in 'span' to start with, since a span <T,N> should
not have iterators from a span<T,M> be interchangable. This now
provides some level of uniqueness by leveraging the type-system --
something the previous implementation using pointers was unable to
accomplish.