✨ Add conversion support to proxy_iterator
C++ iterators require that non-const iterators be convertible to
const iterators.
The initial implementation of the proxy_iterator did not support this,
which prevented T* from being converted to const T* for pointers,
and also prevented iterators from being converted to const iterators.

This adds the required support to proxy_iterator