🩹 Fix MSVC C++11 check
Microsoft Visual Studios, as always, chooses to be difficult when
dealing with C++ in that it refuses to define __cplusplus as being
201103L, even when working in C++11-compliant mode. This causes
great difficulty for checks that are determining c++ support, since
it is unable to guarantee what mode a compiler is working in.

For the time being, the guard in config.hpp has been removed to allow
MSVC to continue compiling. More work will have to be added later to
allow for the BPSTD_[CPP14|CPP17]_CONSTEXPR variables to be determined
correctly.