🩹 Fix warning for reference-to-temporary
Some versions of GCC with -Wall and -Wextra are compaining about
ternary statements that contain a 'throw' statement that returns a
reference. The warning claims that it's returning a reference to a
temporary -- despite the 'throw' statement actually not returning
anything at all. It's important to note that this warning is not
accurate at all.

The ternary statement is necessary for C++11 constexpr, since
constexpr functions must be only one line.

To fix the false-positive warning, the ternary statement has been
updated so that after the throw statement, it returns a redundant
reference. This reference will never actually be reached, however,
since the throw statement would take effect first -- causing the
ternary to exit.