🩹 Disambiguate invoke calls
It appears that MSVC is choking on invocations to unqualified 'invoke'
calls. The exact reason for this failure is... unclear, but MSVC has
never really been strong at figuring out well-formed C++.

This is an attempt to rectify the issue by fully-qualifying the
invoke calls so that there exists no possible case for ambiguity.