📇 Add custom footer.html for doxygen
Doxygen by default likes to inject the date and time that the
documentation was generated at, alongside the version of the doxygen
generator that was used.

Since a new GitHub Action was added to generate doxygen documentation
with every push to master, this causes diffs with every change -- even
if the change is simply the time that the change was added to Github.

To avoid these perpetual changes to the history, a custom footer has
been added that does not contain the time of generation.