🔨 Add UExtent type name to span implementation
MSVC still chooses not to implement the C++ standard correctly, in
particular with regard to SFINAE.

Valid, conforming code need not rename template arguments that are not
used in a definition that were previously named in a declaration.
However, MSVC chooses to refute this. If you have a named dependent
template argument in a SFINAE expression that is left unnamed in the
defintion of the function, MSVC will choose not to treat it as a valid
SFINAE overload.

The fix to appease Microsoft is to simply give it a name in the
implementation... But this doesn't make it any less annoying.