✨ Add constexpr move and forward utilities
C++11 introduced the std::move and std::forward utilities, which
effectively just translate to casts to different value categories.
This form of cast is implicitly constexpr, however in C++11 the
functions were not marked `constexpr` despite being viable for it.

This omission is detrimental since the lack of constexpr support
causes implementations to either require their own constexpr
equivalents, or for the types to also have to sacrifice constexpr
support.

To aid address this, this introduces constexpr versions of these
functions. They have been implemented in an internal header,
detail/move.hpp, to avoid cyclic dependencies between headers -- but
the primary header which exports the definition will remain utility.hpp
to replicate the C++ api, in which its exported from <utility>

Fixed #4