🚦 Update deployment for doxygen
The doxygen deployment scripts fail if the documentation has not
changed, since git commit will return an error code if there is nothing
staged to be committed.

To fix this, the index is checked first before attempting a commit.
This should allow deployment to succeed when there is nothing to
deploy.