🔨 Move 'config.hpp' to detail
The 'config.hpp' header, which provides information about the current
compiler/os configuration, was previously part of the normal 'include'
directory -- making it part of the public API.

Given that this project aims to project the C++ standard's API wherever
possible, the config file should not be part of the public inclusion
set. As a result, it's been relocated to the 'detail' directory to make
it a private implementation.