🩹 Fix any_cast(const any&) (#6)
The any_cast(const any_cast&) overload was mistakenly throwing a
bad_any_cast on the inverse condition that it should have been. This
would trigger the exception when the cast is valid -- and worse, would
result in a null-dereference when the cast was incorrect.

This condition has been corrected.