📖 Add cppcheck suppression comments
Various places in the code were triggering cppcheck warnings for
code that is correctly conforming to the C++ standard.
In particular, it is triggering over all implicit conversion
constructors -- which are required for the types they are implementing.

It is also incorrectly claiming some static member variables are unused
when they are used just 3 lines below the warning.

Unfortunately, cppcheck does not offer any way to disable these warnings
in an easily suppliable format (unlike just a simple .clang-tidy file),
and so these are being addressed with inline suppression comments.