🩹 Unconditionally define 'is_final'
is_final was previously conditionally defined based on whether the
library was being built for C++14 or not. This leads the library to
be version-locked, which is undesirable.

Instead, is_final is now unconditionally defined -- however it
defaults to false_type unless compiler support is available. This
requires an expicitly specialized list of 'final' types made by the
consumer to be used accurately in C++11 mode.