✨ Add get<T>(tuple) implementation
The C++14 standard introduced "T" overloads for std::get to get elements
from a tuple rather than requiring the index.

This provides implementations of those overloads.

Note: Unlike the standard implementation, these get<T> overloads will
not be ADL-callable from a tuple. This is fundamentally not fixible
without introducing an orthogonal tuple implementation in this library;
which would not be ideal.