✨ Add inline visibility macro
`inline` symbols in C++ will always make an effort to resolve to a
single symbol when used in a multi-library context. This means that
even though this library is header-only, if it's used across library
boundaries, then a linked-in executable will attempt to resolve only
a single implementation of the function.

In many (if not most) cases this is desirable. However, given that this
library is aiming to emulate the C++ standard library -- it would be
better for these symbols to be aggressively inlined, if possible, and
for these symbols to *not* be exported for consideration when resolving
symbols.

To do this, the same approach adopted by libc++ and libstdc++ has been
used: defining a new macro which defines the inline visibiliy.

This new visibility specifies that the symbol should remain hidden,
be inlined if possible, and should not be considered when used with
instrumentation.