✨ Add variant_union implementation utility
std::variant's implementation requires that for constexpr support, that
if a variant's underlying members contain trivial destructors, then the
variant's destructor, too, must be trivial. This necessitates the use of
a union -- since placement new can never be constexpr, and would be
unable to condition call the destructor on the underlying type.

When a C++11 union contains a non-trivial data-member, the destructor
is implicitly deleted and requires the author to define it. However,
by defining it -- it becomes no longer trivial -- meaning we cannot
simply implement it in all cases. This requires a divergence based
entirely on whether the types are trivial.

This has been implemented here in the form of 'variant_union'. This type
is effectively a union itself, recursive, supporting N types. If the
types are all trivial, then the destructor is not explicitly defined
which makes it trivial. This is divergent on a 'bool' template
parameter.

This will be an implementation detail in the Backport's variant.