🔨 Add workaround to BPSTD_INLINE_VISIBILITY
When building 'variant' using 'clang' and 'clang-cl' that uses MSVC
codegen, it results a compiler code generation failure due to the
'always_inline' attribute. This attribute is only used as a means to
attempt to control access to the otherwise internal symbols anyway,
so for these builds this has been disabled.

Symbols are still documented, however, with the internal visibility
attribute to ensure that these symbols do not become part of the
library's ABI.