🩹 Fix any_cast with reference types
The `any_cast` is currently broken with any T type where T is a
possibly CV-qualified reference type. The standard requires that this
return a proper reference to the underlying stored value.

The reason for this failure is due to it calling `any_cast(any*)` in
the implementation by just forwarding along the `T` type. Since the
pointer-overload for the `any_cast` does not work with CV-qualified
reference types, the code fails to compile and does not generate the
correct output.

To fix this, the implementation of `bpstd::any_cast` simply removes the
cvref-ness of the input prior to propagating to the `any_cast`