✨ Add SFINAE to underlying_type
When std::underlying_type was introduced into C++11, it simply
left the primary template undefined in the case that T was not an
enum type. C++17 changes this to be defined, but without any `type`
member -- which allows for types to use this as a SFINAE condition.

This updates the `underlying_type` implementation to do the same.