✨ Enable /W4 and /WX on MSVC
The compile flags for building with MSVC previously only used CMake's
builtin `/W3` flag (which occurs on older versions of CMake). This does
not cover all info / warnings that might be generated by MSVC which may
trip clients of this library who are building with this level.

To avoid potential issues, this project has been updated to build
against both `/W4` (the highest warning level), and `/WX` -- which
compiles warnings as errors.