✨ Add SourceGroup CMake module
Some CMake generator targets don't preserve directory structure for
source-files, which leads to a coallescing of subsystems into a global
scope. Both MSVC and Xcode do this, which is causing `detail` directory
headers to appear as though they are publicly includable headers.

A new module has been added which preserves the directory structure,
and the project has been updated to make use of this.