🔨 Suppress MSVC "deprecated" warning for copy
MSVC willfully chooses to be non-standard and warns on the use of
`std::copy` with an unchecked pointer in `string_view`'s implementation.

The C++ standard does _not_ deprecate such a use, and such a use is not
invalid to be done, but MSVC -- an intellectual -- chooses to deprecate
this function.

There doesn't appear to be an easy way to prevent this warning from
seeping through, since even using the `#pragma warning(disable:...)`
does not seem to fix the issue -- so the logic has been modified to be
written inline without the use of `std::copy`