🚦 Split CI to separate workflows
The `main.yml` Github Workflow has been broken ever since Github
deprecated and removed support for the `::set-env` syntax. Unfortunately
this has lead to failures in the repo due simply to Github disallowing
the old (insecure) syntax.

This now splits the workflows so that it works each of the
various operating systems operate on an independent workflow.

Additionally, workflows are now only triggered when any code relating
to the workflows has changed.