🔨 Merge coverage data after completion on CI
The code-coverage Github Workflow is currently uploading two competing
coverage tracefiles of the same data to Coveralls. The expectation was
that Coveralls merges this data; the reality is that Coveralls just
erases the first one and keeps the second one, which introduces a
race-condition.

To fix this, all coverage data is instead uploaded as an artifact,
and then downloaded in a later pass to be merged together before
uploading it to Coveralls. This now ensures that different compilers
can each be tested to be able to test different code-paths.